requirements.txt
# sudo apt-get install libleptonica-dev 
# sudo apt-get install tesseract-ocr tesseract-ocr-dev
# sudo apt-get install libtesseract-dev
# sudo apt install tesseract-ocr

# https://ecsocman.hse.ru/data/2010/05/26/1212617593/Doklad-Pages-001-392-posle-obreza-170x240mm.pdf

# http://exploringafrica.matrix.msu.edu/wp-content/uploads/2015/06/Economic-data-table-pt.-1.png
# http://exploringafrica.matrix.msu.edu/wp-content/uploads/2015/06/Economic-data-table-pt-2.png
# http://exploringafrica.matrix.msu.edu/wp-content/uploads/2016/02/Table-7-pt.-3.png
# https://illinoisepi.org/site/wp-content/uploads/economic-table.jpg
# https://d3i71xaburhd42.cloudfront.net/feedf5b8040023d71b9f90cfddb8775fb16ff078/27-Table1-1.png
# https://rus.azattyq-ruhy.kz/userfiles/%D1%8D%D0%BA%D0%BE%D0%BB%D0%BE%D0%B3%D0%B8%D1%8F/c0bb2f80-f8ad-46d8-bdbd-08f8ef672f7e.jpg
# https://cf.ppt-online.org/files/slide/d/dgbFzC08V4m5usoPBvipOU2ZjRYWLcktDHflaq/slide-1.jpg

# https://aviapoisk.kz/raspisanie/aeroporta/ustkamenogorsk
# https://www.gks.ru/bgd/regl/b09_47/IssWWW.exe/Stg/2-13.htm

dependency-injector
flask
pandas
lxml
flask-socketio==4.3.1
python-engineio==3.13.2
python-socketio==4.6.0
websocket
flask-login
pymongo
html5lib
BeautifulSoup4
simplejson
python-interface
tabula-py
opencv-python
table_ocr
wikipedia
camelot-py
SPARQLWrapper
mediatr
asyncio
redis
rake-nltk
git+https://github.com/Koziev/ruword2tags
git+https://github.com/Koziev/rutokenizer
git+https://github.com/Koziev/rupostagger
git+https://github.com/Koziev/rulemma
git+https://github.com/Koziev/rusyllab
python-crfsuite
pathlib
python-rake
langdetect
flasgger
stanfordnlp==0.2.0
networkx==2.3
textacy
spacy
spacy-udpipe
spacy[transformers]
